agents/agent-frontend/requirements.txt
flask==2.0.1
werkzeug==2.0.1
azure-storage-blob==12.9.0
python-dotenv==0.19.1
markdown